Update to upstream scamper-cvs-20230224
https://mailman.caida.org/pipermail/scamper-announce/2023-February/000029.html

+++ scamper/scamper_writebuf.c
@@ -1,12 +1,12 @@
 /*
  * scamper_writebuf.c: use in combination with select to send without blocking
  *
- * $Id: scamper_writebuf.c,v 1.47 2020/03/17 07:32:16 mjl Exp $
+ * $Id: scamper_writebuf.c,v 1.48 2022/12/09 09:37:42 mjl Exp $
  *
  * Copyright (C) 2004-2006 Matthew Luckie
  * Copyright (C) 2006-2010 The University of Waikato
  * Copyright (C) 2014      The Regents of the University of California
- * Copyright (C) 2014-2016 Matthew Luckie
+ * Copyright (C) 2014-2022 Matthew Luckie
  * Author: Matthew Luckie
  *
  * This program is free software; you can redistribute it and/or modify
@@ -272,11 +272,11 @@ size_t scamper_writebuf_len2(const scamper_writebuf_t *wb,char *str,size_t len)
       c++;
     }
 
-  string_concat(str, len, &off, "%d,%d%s", k, c, (k != 0) ? ":" : "");
+  string_concat(str, len, &off, "%d,%d%s", (int)k, c, (k != 0) ? ":" : "");
   for(node=slist_head_node(wb->iovs); node != NULL; node=slist_node_next(node))
     {
       iov = slist_node_item(node);
-      string_concat(str, len, &off, " %d", iov->iov_len);
+      string_concat(str, len, &off, " %d", (int)iov->iov_len);
     }
 
   return k;
